Add testing

+++ bitrix_cleaner.go
@@ -14,17 +14,24 @@ import (
 	"time"
 )
 
+var all = false
+var test = false
+var done chan struct{}
+
+
 func main() {
 
 	runtime.GOMAXPROCS(runtime.NumCPU())
 
 	path, err := os.Getwd()
-	all := false
+	all = false
+	test = false
 	dirs := []string{"managed_cache", "stack_cache", "cache", "html_pages"}
-	done := make(chan struct{}, len(dirs))
+	done = make(chan struct{}, len(dirs))
 
 	flag.StringVar(&path, "path", path, "Path to bitrix root")
-	flag.BoolVar(&all, "all", false, "Process all files (if not provided then the expired files will be processed only)")
+	flag.BoolVar(&all, "all", all, "Process all files (if not provided then the expired files will be processed only)")
+	flag.BoolVar(&test, "donotremove", test, "Do not remove files. Run for testing.")
 	flag.Parse()
 	path = strings.TrimSuffix(path, string(os.PathSeparator))
 
@@ -43,7 +50,7 @@ func main() {
 		fmt.Fprint(os.Stderr, path+string(os.PathSeparator)+"bitrix is not directory.")
 	} else {
 		for _, dir := range dirs {
-			go processDir(path+string(os.PathSeparator)+"bitrix"+string(os.PathSeparator)+dir, all, done)
+			go processDir(path+string(os.PathSeparator)+"bitrix"+string(os.PathSeparator)+dir)
 		}
 		waitUntil(done, len(dirs))
 	}
@@ -55,7 +62,7 @@ func waitUntil(done <-chan struct{}, len int) {
 	}
 }
 
-func processDir(dir string, all bool, done chan<- struct{}) {
+func processDir(dir string) {
 	_, err := os.Stat(dir)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, "Error processing "+dir)
@@ -78,9 +85,13 @@ func processDir(dir string, all bool, done chan<- struct{}) {
 
 func processFiles(path string, info os.FileInfo, err error) error {
 	if err == nil && !info.IsDir() && strings.HasSuffix(path, ".php") {
-		err := os.Remove(path)
-		if err != nil {
-			return err
+		if test {
+			fmt.Println("Removing "+path)
+		} else {
+			err := os.Remove(path)
+			if err != nil {
+				return err
+			}
 		}
 	}
 	return err
@@ -114,9 +125,13 @@ func processExpiredFiles(path string, info os.FileInfo, err error) error {
 				tm, err := strconv.Atoi(match[1])
 				if err == nil {
 					if int64(tm) < tmnow {
-						err := os.Remove(path)
-						if err != nil {
-							return err
+						if test {
+							fmt.Println("Removing "+path)
+						} else {
+							err := os.Remove(path)
+							if err != nil {
+								return err
+							}
 						}
 					}
 				}
